% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRED_MD.R
\docType{data}
\name{FRED_MD}
\alias{FRED_MD}
\title{Federal Reserve Bank at St Louis.}
\format{
An object of class \code{"data.frame"}.
}
\source{
\href{https://www.stlouisfed.org/}{https://www.stlouisfed.org/}
}
\usage{
data("FRED_MD")
}
\description{
Data obtained from the Federal Research Bank after process
to remove missing values.
}
\keyword{datasets}
